# swagger.yaml
openapi: 3.0.0
info:
  title: AUREA Backend API
  description: |
    Comprehensive Portfolio Management Platform API with 25 professional endpoints.
    
    **Features:**
    - JWT Authentication System
    - Complete User Profile Management
    - Complete Portfolio CRUD Operations
    - One-Click Vercel Deployment with Smart HTML Generation
    - Dynamic Responsive HTML from User Data
    - Case Study Management with Intelligent Transformation
    - Professional Image Upload (Cloudinary)
    - Publishing System with Slug Management
    - View Tracking & Analytics
    
    **Case Study Smart Transformation (October 2025):**
    - Automatically detects real vs template/default content
    - Uses database data when available, professional fallbacks when empty
    - Generates fully responsive mobile-optimized HTML pages
    - Auto-marks projects with hasCaseStudy flags for portfolio integration
    - Comprehensive test suite for verification
    
    **Total Endpoints: 32**
    - Authentication: 3 endpoints
    - User Management: 7 endpoints
    - Portfolio Management: 8 endpoints
    - Site Publishing & Deployment: 3 endpoints
    - Case Study Management: 4 endpoints
    - Image Upload: 2 endpoints
    - PDF Export: 5 endpoints
  version: 1.0.0
  contact:
    name: AUREA Development Team
  license:
    name: ISC
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.aurea.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information
paths:
...
  /api/pdf/portfolio/{portfolioId}:
    get:
      summary: Export portfolio as PDF
      description: |
        Export a portfolio as PDF document with inline viewing.

        **Features:**
        - Preserves exact styling from templateConvert.js
        - Supports custom page formats and orientations
        - Optional authentication for public portfolios
        - Can export main portfolio or specific case studies

        **Access Control:**
        - Owner: Full access with optional save to filesystem
        - Public: Access only if portfolio is published
      tags:
        - PDF Export
      security:
        - bearerAuth: []
        - {}
      parameters:
        - in: path
          name: portfolioId
          required: true
          schema:
            type: string
          description: Portfolio ID to export
          example: "507f1f77bcf86cd799439011"
        - in: query
          name: pageType
          schema:
            type: string
            default: portfolio
          description: |
            Type of content to export:
            - 'portfolio': Main portfolio page only (default)
            - 'all': All pages including case studies
            - 'case-study-{id}': Specific case study by project ID
          example: "portfolio"
        - in: query
          name: save
          schema:
            type: string
            enum: [true, false]
            default: false
          description: Save to filesystem (owner only, for debugging)
        - in: query
          name: format
          schema:
            type: string
            enum: [A4, A3, Letter, Legal]
            default: A4
          description: PDF page format
        - in: query
          name: landscape
          schema:
            type: string
            enum: [true, false]
            default: false
          description: Use landscape orientation
      responses:
        '200':
          description: PDF generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
                example: application/pdf
            Content-Disposition:
              schema:
                type: string
                example: inline; filename="portfolio-export.pdf"
            Content-Length:
              schema:
                type: integer
                example: 524288
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied - Portfolio not published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: PDF generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pdf/portfolio/{portfolioId}/complete:
    get:
      summary: Export complete portfolio with case studies
      description: |
        Export portfolio and all associated case studies as a single combined PDF.

        **Features:**
        - Combines main portfolio and all case studies
        - Maintains proper page ordering and navigation
        - Preserves all styling and formatting
        - Ideal for complete portfolio documentation
      tags:
        - PDF Export
      security:
        - bearerAuth: []
        - {}
      parameters:
        - in: path
          name: portfolioId
          required: true
          schema:
            type: string
          description: Portfolio ID to export
          example: "507f1f77bcf86cd799439011"
        - in: query
          name: save
          schema:
            type: string
            enum: [true, false]
            default: false
          description: Save to filesystem (owner only)
        - in: query
          name: format
          schema:
            type: string
            enum: [A4, A3, Letter, Legal]
            default: A4
          description: PDF page format
        - in: query
          name: landscape
          schema:
            type: string
            enum: [true, false]
            default: false
          description: Use landscape orientation
      responses:
        '200':
          description: Complete PDF generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
                example: application/pdf
            Content-Disposition:
              schema:
                type: string
                example: inline; filename="portfolio-complete.pdf"
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: PDF generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pdf/portfolio/{portfolioId}/download:
    get:
      summary: Download portfolio PDF
      description: |
        Download portfolio as PDF file (forces download dialog).

        **Features:**
        - Forces browser download instead of inline viewing
        - Same export options as inline endpoint
        - Useful for saving PDFs directly to user's device
      tags:
        - PDF Export
      security:
        - bearerAuth: []
        - {}
      parameters:
        - in: path
          name: portfolioId
          required: true
          schema:
            type: string
          description: Portfolio ID to download
          example: "507f1f77bcf86cd799439011"
        - in: query
          name: pageType
          schema:
            type: string
            default: portfolio
          description: Type of content to export
        - in: query
          name: format
          schema:
            type: string
            enum: [A4, A3, Letter, Legal]
            default: A4
          description: PDF page format
        - in: query
          name: landscape
          schema:
            type: string
            enum: [true, false]
            default: false
          description: Use landscape orientation
      responses:
        '200':
          description: PDF ready for download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
                example: application/pdf
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="portfolio-export.pdf"
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...